document_processor/requirements-docprocessor.txt
# Document Processor Requirements для PDF Converter Pipeline v4.0
# Зависимости для обработки PDF документов с множественными движками

# ================================================================================================
# DOCLING - Основной движок для layout parsing и извлечения контента
# ================================================================================================
docling==2.45.0

# ================================================================================================
# OPTICAL CHARACTER RECOGNITION (OCR)
# ================================================================================================
# PaddleOCR для китайского языка
paddleocr>=3.2.0
paddlepaddle>=3.1.0

# Tesseract Python биндинги  
pytesseract>=0.3.10

# ================================================================================================  
# ИЗВЛЕЧЕНИЕ ТАБЛИЦ
# ================================================================================================
# Tabula-Py для извлечения таблиц из PDF
tabula-py>=2.9.0

# Camelot для сложных таблиц (альтернатива)
camelot-py[cv]>=0.11.0

# ================================================================================================
# ОБРАБОТКА ИЗОБРАЖЕНИЙ И КОМПЬЮТЕРНОЕ ЗРЕНИЕ
# ================================================================================================
opencv-python>=4.8.1
opencv-contrib-python>=4.8.1
scikit-image>=0.21.0
Pillow>=10.1.0
imageio>=2.33.0

# ================================================================================================
# PDF И ДОКУМЕНТЫ
# ================================================================================================
PyPDF2>=3.0.1
PyMuPDF>=1.23.0
pdfplumber>=0.10.3
pymupdf>=1.23.0
pdf2image>=1.16.3

# ================================================================================================
# МАШИННОЕ ОБУЧЕНИЕ И AI
# ================================================================================================
# Transformers для работы с моделями
transformers>=4.35.0
torch>=2.1.0
torchvision>=0.16.0

# Hugging Face datasets и utilities
datasets>=2.14.0
accelerate>=0.24.0

# Для работы с ONNX моделями
onnxruntime>=1.16.0

# ================================================================================================
# ВЕБЕРНЫЙ СЕРВЕР И API
# ================================================================================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# HTTP клиенты
httpx>=0.25.0
aiohttp>=3.9.0
requests>=2.31.0

# ================================================================================================
# ОБРАБОТКА ДАННЫХ
# ================================================================================================
numpy>=1.24.0
pandas>=2.1.0
scipy>=1.11.0

# JSON обработка
orjson>=3.9.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ================================================================================================
# СТРУКТУРИРОВАННЫЕ ДАННЫЕ И ФОРМАТЫ
# ================================================================================================
# Markdown обработка
markdown>=3.5.0
python-markdown-math>=0.8.0

# YAML и конфигурация
pyyaml>=6.0.1
python-dotenv>=1.0.0

# XML обработка
lxml>=4.9.3
beautifulsoup4>=4.12.0

# ================================================================================================
# УТИЛИТЫ ДЛЯ ТЕКСТА И NLP
# ================================================================================================
# Китайский язык
jieba>=0.42.1
pypinyin>=0.51.0

# Общие NLP утилиты  
regex>=2023.10.3
nltk>=3.8.1
spacy>=3.7.0

# Детекция языка
langdetect>=1.0.9
polyglot>=16.7.4

# ================================================================================================
# МОНИТОРИНГ И ЛОГИРОВАНИЕ
# ================================================================================================
structlog>=23.2.0
prometheus-client>=0.19.0
psutil>=5.9.0

# ================================================================================================
# ФАЙЛОВАЯ СИСТЕМА И I/O
# ================================================================================================
pathlib2>=2.3.7
watchdog>=3.0.0
filelock>=3.13.0

# Архивы
zipfile-deflate64>=0.2.0

# ================================================================================================
# ДАТА И ВРЕМЯ
# ================================================================================================
python-dateutil>=2.8.0
pytz>=2023.3

# ================================================================================================
# ПАРАЛЛЕЛИЗМ И ASYNC
# ================================================================================================
asyncio-mqtt>=0.16.0
aiofiles>=23.2.0

# Многопроцессорность
multiprocessing-logging>=0.3.4

# ================================================================================================
# DEBUGGING И DEVELOPMENT
# ================================================================================================
ipython>=8.17.0
jupyter>=1.0.0

# Профилирование
memory-profiler>=0.61.0
line-profiler>=4.1.0

# ================================================================================================
# СОВМЕСТИМОСТЬ И СТАБИЛЬНОСТЬ
# ================================================================================================
packaging>=23.2
setuptools>=68.0.0
wheel>=0.41.0
typing-extensions>=4.8.0

# ================================================================================================
# СПЕЦИФИЧЕСКИЕ УТИЛИТЫ ДЛЯ PDF PIPELINE
# ================================================================================================
# Извлечение метаданных
exifread>=3.0.0
python-magic>=0.4.27

# Графики и диаграммы в документах  
matplotlib>=3.8.0
seaborn>=0.13.0

# Геометрия для layout analysis
shapely>=2.0.0
matplotlib-scalebar>=0.8.1